refactor: API 엔티티 마이그레이션 정리
