+++ src/pages/templates/model/template-list.ts
@@ -1,4 +1,5 @@
-import { type TemplateSummary, getRelativeTimeLabel } from "@/shared";
+import type { TemplateSummary } from "@/entities/template";
+import { getRelativeTimeLabel } from "@/shared";
 
 export const getRelativeCreatedLabel = (createdAt: string) =>
   getRelativeTimeLabel(createdAt, {
