--- src/shared/types/api-models.type.ts
@@ -1,11 +1,0 @@
-export type {
-  ChoiceBranchConfig,
-  ChoiceFollowUp,
-  ChoiceOption,
-  ChoiceResponse,
-  NodeAddRequest,
-  NodeDefinitionResponse,
-  NodeSelectionResult,
-  WorkflowResponse,
-} from "../../entities/workflow";
-export type { TemplateDetail, TemplateSummary } from "../../entities/template";
