+++ src/shared/types/index.ts
@@ -1,2 +1,1 @@
 export * from "./api.type";
-export * from "./api-models.type";
